+++ src/core_modules/capture-core/components/Pages/common/EnrollmentOverviewDomain/EnrollmentPageLayout/index.js
@@ -1,0 +1,9 @@
+// @flow
+
+export {
+    DefaultWidgetsForEnrollmentOverview,
+} from './DefaultEnrollmentLayout.constants';
+export { DataStoreKeyByPage } from './DefaultEnrollmentLayout.constants';
+export { WidgetTypes } from './DefaultEnrollmentLayout.constants';
+export * from './LayoutComponentConfig';
+export { EnrollmentPageLayout } from './EnrollmentPageLayout';