+++ src/core_modules/capture-core/components/WidgetAssignee/hooks/index.js
@@ -1,0 +1,2 @@
+// @flow
+export { useUserAvatar } from './useUserAvatar';