+++ src/core_modules/capture-core/components/WidgetErrorAndWarning/WidgetError/WidgetError.types.js
@@ -3,5 +3,9 @@ import type { Message } from '../content/WidgetErrorAndWarningContent.types';
 
 export type Props = {|
     error?: Array<Message>,
+|}
+
+export type PlainProps = {|
+    error?: Array<Message>,
     ...CssClasses
 |}