+++ src/core_modules/capture-core/components/WidgetEventSchedule/Assignee/Assignee.types.js
@@ -4,4 +4,5 @@ import type { UserFormField } from '../../FormFields/UserField';
 export type Props = {
     ...CssClasses,
     assignee?: UserFormField,
+    onSetAssignee: (user: UserFormField) => void,
 };