+++ src/core_modules/capture-core/components/WidgetProfile/DataEntry/helpers/index.js
@@ -15,3 +15,4 @@ export {
 export { GEOMETRY, getFeatureType, getDataElement, getLabel } from './geometry';
 export { convertClientToView } from './convertClientToView';
 export { escapeString } from './escapeString';
+export { handleAPIResponse, REQUESTED_ENTITIES } from './handleAPIResponse';