+++ src/core_modules/capture-core/metaData/FormFieldPluginConfig/FormFieldPluginConfig.js
@@ -4,13 +4,14 @@
 /* eslint-disable no-restricted-syntax */
 
 import isFunction from 'd2-utilizr/lib/isFunction';
-import type { PluginFormFieldMetadata } from '../../components/D2Form/FormFieldPlugin';
+import { DataElement } from '../DataElement';
 
 export class FormFieldPluginConfig {
     _id: string;
     _name: string;
     _pluginSource: string;
-    _fields: Map<string, PluginFormFieldMetadata>;
+    _fields: Map<string, DataElement>;
+    _customAttributes: Map<string, { IdFromPlugin: string, IdFromApp: string }>;
 
     constructor(initFn: ?(_this: FormFieldPluginConfig) => void) {
         initFn && isFunction(initFn) && initFn(this);
@@ -32,11 +33,11 @@ export class FormFieldPluginConfig {
         this._name = value;
     }
 
-    get fields(): Map<string, PluginFormFieldMetadata> {
+    get fields(): Map<string, DataElement> {
         return this._fields;
     }
 
-    set fields(value: Map<string, PluginFormFieldMetadata>) {
+    set fields(value: Map<string, DataElement>) {
         this._fields = value;
     }
 
@@ -44,11 +45,19 @@ export class FormFieldPluginConfig {
         return this._pluginSource;
     }
 
+    get customAttributes(): Map<string, { IdFromPlugin: string, IdFromApp: string }> {
+        return this._customAttributes;
+    }
+
+    set customAttributes(value: Map<string, { IdFromPlugin: string, IdFromApp: string }>) {
+        this._customAttributes = value;
+    }
+
     set pluginSource(value: string) {
         this._pluginSource = value;
     }
 
-    addField(idFromPlugin: string, field: PluginFormFieldMetadata) {
+    addField(idFromPlugin: string, field: DataElement) {
         if (!this.fields.has(idFromPlugin)) {
             this.fields.set(idFromPlugin, field);
         }