+++ src/core_modules/capture-core/utils/api/index.js
@@ -1,3 +1,4 @@
 // @flow
 export { makeQuerySingleResource } from './makeQuerySingleResource';
+export { handleAPIResponse, REQUESTED_ENTITIES } from './handleAPIResponse';
 export type * from './api.types';